Новый компонент FavoritePlaceCard
