+++ src/components/favorite-location-list/favorite-location-list.jsx
@@ -1,0 +1,42 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import FavoritePlaceCard from '../favorite-place-card/favorite-place-card';
+
+const FavoriteLocationList = ({city, cityFavoriteOffers}) => {
+
+  const cityFavoriteOffersElements = cityFavoriteOffers.map((offer, index) => {
+    return (
+      <FavoritePlaceCard favoriteOffer={offer} key={index} />
+    );
+  });
+
+  return (
+    <li className="favorites__locations-items">
+      <div className="favorites__locations locations locations--current">
+        <div className="locations__item">
+          <a className="locations__item-link" href="#">
+            <span>{city}</span>
+          </a>
+        </div>
+      </div>
+      <div className="favorites__places">
+        {cityFavoriteOffersElements}
+      </div>
+    </li>
+  );
+};
+
+FavoriteLocationList.propTypes = {
+  city: PropTypes.string.isRequired,
+  cityFavoriteOffers: PropTypes.arrayOf(PropTypes.shape({
+    images: PropTypes.array.isRequired,
+    accommodation: PropTypes.shape({
+      price: PropTypes.string.isRequired,
+      rating: PropTypes.number.isRequired,
+      title: PropTypes.string.isRequired,
+      type: PropTypes.string.isRequired,
+    }).isRequired,
+  })).isRequired
+};
+
+export default FavoriteLocationList;
